pyproject.toml
[project]
name = "ai-agent-perpexity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "ipython>=9.6.0",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "markdown>=3.5.0",
    "matplotlib>=3.10.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.12",
    "weasyprint>=62.0",
]
